signature_generator/requirements.txt
Flask==2.3.2
Jinja2==3.1.2
Werkzeug==2.3.3
MarkupSafe==2.1.3
itsdangerous==2.1.2
Pillow==9.5.0